+++ src/Parser/JSON/TestHook/AbstractTestHook.php
@@ -15,6 +15,7 @@ abstract class AbstractTestHook
 
     public function __construct()
     {
+        /** @psalm-suppress DocblockTypeContradiction */
         if (null === self::$logFile) {
             $file = fopen($this->getLogFilename(), 'wt');
             if (! \is_resource($file)) {