+++ src/Runner/Runner.php
@@ -46,6 +46,11 @@ class Runner implements EventSubscriberInterface
         $this->processFactory = $processFactory;
         $this->filter = $filter;
         $this->pipelineCollection = $pipelineCollection;
+        /**
+         * @psalm-suppress MixedPropertyTypeCoercion
+         *
+         * @see https://github.com/vimeo/psalm/issues/8103
+         */
         $this->queuedProcesses = new \SplQueue();
         $this->exitCode = 0;
     }