+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file. For previou
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ## Unreleased
- * ...
+### Added
+ * Add support for Symfony 6 [#168](https://github.com/facile-it/paraunit/pull/168)
+### Removed
+ * Drop support for Symfony < 4.4 [#168](https://github.com/facile-it/paraunit/pull/168)
 
 ## [1.2.1] - 2021-03-25
 ### Added